+++ CHANGELOG.md
@@ -8,8 +8,11 @@
 - [x]  Texture Loading
     - [x]  Offset
     - [x]  Load Surface, then Texture
+- [x]  Preliminary Shader Support
 - [x]  (Other) Create Versioning
 - [x]  (Other) Move to Lume
+#### Changed
+- [x]  Revamped Graphics Renderer
 
 ---